e2e/requirements.e2e.txt
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
kafka-python==2.0.2
pymongo==4.6.1
redis==5.0.1
psycopg2-binary==2.9.9
tenacity==8.2.3  # Para retry em operações que podem falhar temporariamente
